Move code and wrap only if compatible
