+++ .github/copilot-instructions.md
@@ -1,0 +1,7 @@
+All code MUST ALWAYS follow the `naming-things` guidelines!
+
+Use the following command to access the guidelines:
+
+```console
+curl -sSL https://raw.githubusercontent.com/codingjoe/naming-things/refs/heads/main/README.md | cat
+```
